TG-30: Solving comments
