+++ src/Components/Admins/admins.module.css
@@ -67,6 +67,7 @@
 }
 
 .buttonUpdate:hover,
-.buttonAddAdmin:hover {
+.buttonAddAdmin:hover,
+.row:hover {
   cursor: pointer;
 }
